Created a separate file Wallet.java for Wallet API calls.
